requirements.txt
streamlit
streamlit-chat
requests
fastapi
uvicorn
python-dotenv
anthropic
